refactor: move withSSRData to services/initial-ssr-data
