+++ src/server/utils.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { Switch, Route } from 'react-router-dom';
 
 import { getEnv } from 'src/utils/env';
-
-import withSSRData from '../shared-components/withSSRData';
+import withSSRData from 'src/services/initial-ssr-data/withSSRData';
 
 // Due to usage of HOCs, we have to traverse the component tree to find the
 // final wrapped component which contains the static server-side methods
@@ -28,10 +27,11 @@ export function findFinalComponent(component) {
 
 export function getComponent(component, routeRenderer) {
   // because route renderers are a wrapper around the actual component,
-  // they return the element created by react and not the actual component so we 
+  // they return the element created by react and not the actual component so we
   // need to point to the actual component which is stored in the 'type' property
   // in the returned react element
   const reactElement = routeRenderer ? routeRenderer() : null;
+
   return reactElement ? reactElement.type : component;
 }
 
@@ -57,10 +57,10 @@ export function renderRoutes(routes, extraProps = {}, switchProps = {}) {
 
               return (needsWithSSRData)
                 ? React.createElement(withSSRData(component), componentProps)
-                : React.createElement(component, componentProps)
+                : React.createElement(component, componentProps);
             }}
           />
-        )
+        );
       })}
     </Switch>
   );
